Switch to using a single file for the chapter
The split approach is annoying because it does not automatically
rerender the preview on change. It also is more annoying to manage,
since the notebooks that share the names of the sections clutter the
results of file lookup.

Moreover, I learned to embrace the folding system in Neovim. It solves
the problem of navigating large files when they are split into sections,
like here.